+++ lib/labeled.dart
@@ -22,14 +22,13 @@ class LabeledStyle extends InheritedTheme {
   const LabeledStyle.fallback({Key key})
       : format = LabeledFormat.above,
         spacing = 0,
-        labelStyle = const TextStyle(fontSize: 10),
+        labelStyle = const TextStyle(fontSize: 14),
         crossAxisAlignment = CrossAxisAlignment.start,
         mainAxisAlignment = MainAxisAlignment.start,
         super(key: key, child: null);
 
   static LabeledStyle of(BuildContext context) {
-    return context.dependOnInheritedWidgetOfExactType<LabeledStyle>() ??
-        const LabeledStyle.fallback();
+    return context.dependOnInheritedWidgetOfExactType<LabeledStyle>() ?? const LabeledStyle.fallback();
   }
 
   @override
@@ -42,8 +41,7 @@ class LabeledStyle extends InheritedTheme {
 
   @override
   Widget wrap(BuildContext context, Widget child) {
-    final LabeledStyle defaultTextStyle =
-        context.findAncestorWidgetOfExactType<LabeledStyle>();
+    final LabeledStyle defaultTextStyle = context.findAncestorWidgetOfExactType<LabeledStyle>();
     return identical(this, defaultTextStyle)
         ? child
         : LabeledStyle(
@@ -67,21 +65,21 @@ class Labeled extends StatelessWidget {
   final TextStyle labelStyle;
   final CrossAxisAlignment crossAxisAlignment;
   final MainAxisAlignment mainAxisAlignment;
+  final EdgeInsetsGeometry labelMargin;
 
-  const Labeled(
-      this.label,
+  const Labeled(this.label,
       {Key key,
-      this.child,
+      @required this.child,
       this.format,
       this.labelStyle,
       this.spacing,
       this.crossAxisAlignment,
+      this.labelMargin,
       this.mainAxisAlignment})
       : super(key: key);
 
-  Widget _spacing(LabeledFormat format) => format == LabeledFormat.above
-      ? SizedBox(height: spacing)
-      : SizedBox(width: spacing);
+  Widget _spacing(LabeledFormat format, double spacing) =>
+      format == LabeledFormat.above ? SizedBox(height: spacing) : SizedBox(width: spacing);
 
   @override
   Widget build(BuildContext context) {
@@ -90,15 +88,17 @@ class Labeled extends StatelessWidget {
     var spacing = this.spacing ?? style.spacing ?? 0;
 
     List<Widget> content = [
-      Text(label, style: this.labelStyle ?? style.labelStyle),
+      Padding(
+        padding: labelMargin ?? EdgeInsets.zero,
+        child: Text(label, style: this.labelStyle ?? style.labelStyle),
+      ),
       if (format == LabeledFormat.across) Spacer(),
-      if (format != LabeledFormat.across && spacing > 0) _spacing(format),
+      if (format != LabeledFormat.across && spacing > 0) _spacing(format, spacing),
       child
     ];
 
     return Flex(
-        direction:
-            format == LabeledFormat.above ? Axis.vertical : Axis.horizontal,
+        direction: format == LabeledFormat.above ? Axis.vertical : Axis.horizontal,
         crossAxisAlignment: crossAxisAlignment ?? style.crossAxisAlignment ?? CrossAxisAlignment.start,
         mainAxisAlignment: mainAxisAlignment ?? style.mainAxisAlignment ?? MainAxisAlignment.start,
         children: content);