+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## [0.0.1] - TODO: Add release date.
+## [1.0.0] 
+ - Added label margin
 
-* TODO: Describe initial release.
+## [0.0.2+1] 
+ - Fixed layoutstyle spacing
+
+## [0.0.2] 
+ - Increased default label font size.
+
+## [0.0.1] 
+ - initial release.